feat: fix broken medical datasets (#39)
We currently have 3 medical datasets that aren't working:
1. NIH-Chest-X-ray,
2. skin-cancer,
3. chest-xray-classification.

The fixes in this PR introduce the concept of `target feature` when
handling HF datasets, to explicitly mention the column we would like to
use as the `label` column (that may not exist in the dataset). Also, I
added some handlers and conversions for out-of-the-standard content in
such column.

Datasets (2) and (3) are downloaded successfully and are compatible with
the existing models, but dataset (1) still has an incompatibility issue
when building the embeddings. We should assert if we want dataset (1) at
all in this phase as its data alone occupies 85 Gb.

As part of the testing process, I found a blocker where the disk space
ran out because the cache for datasets wasn't set in a sole place for
all users and runs which ended up filling a drive with small capacity in
the VM. Now, by default all dataset related content is stored in the
`.cache` dir in the project's root.