feat: add text embeddings (#46)
Now text embeddings generation is supported on every available model.
It's a working v1, as the text template for building the class sentences
is currently hardcoded in the dataset class.

Also, I eliminated the `SiglipModel` class as it is a Hugging Face
model, so the code was a 99% duplicate. For supporting the required
tokenizer for the text embeddings in all HF models, I used `AutoModel`,
`AutoTokenizer` and `AutoProcessor` in the `ClosedCLIPModel` class, so
we don't have to explicitly state the tokenizer that correspond to each
model type (as the repo contains several model variations already).

Went ahead with some standardisation of tensor shapes in the output of
methods that appear in the model classes and when to convert the tensors
from GPU to CPU, and vice versa.