feat: detach resources from providers (#53)
Types and definitions of models and datasets are now dynamically fetched
from a `sources` folder. With this change, previously built-in resources
can be modified, silenced or eliminated if the user wants to. Also,
enables seamless addition of new models and datasets by just adding the
corresponding implementation (new type) or definition (instance of an
existing type).

On the side, refactored the naming convention of several components that
were moved with the previous changes, so they share the same patterns.
E.g. the model base class went from `CLIPModel` to `Model` as the class
itself is related to a higher abstraction than the one of clip.

Also, added the singleton pattern and some minor changes to the dataset
and model providers in order to avoid partially initialised module
errors that appeared when dynamic resources were (tried) loaded before
all local resources were successfully imported. The singleton pattern is
specifically used to avoid the creation of several provider instances as
the original global provider couldn't be loaded on initialisation.