chore: shuffle data on the generated HF splits
Generated splits used the data in the input order, which caused skewed datasets, e.g. containing only one class because the last 15% of the train dataset only had samples from one class. Shuffling the data when generating the splits is enough to prevent this dependency from the input order in Hugging Face datasets.