chore: shuffle data on the generated HF splits (#74)
Generated splits used the data in the input order, which caused skewed
datasets, e.g. containing only one class because the last 15% of the
train dataset only had samples from that class.

Shuffling the data when generating the splits is enough to prevent the
dependency from the input order when managing Hugging Face datasets.